+++ bridge-svm/programs/bridge-svm/src/instructions/mod.rs
@@ -1,0 +1,7 @@
+pub mod admin;
+pub mod lock;
+pub mod unlock;
+
+pub use admin::*;
+pub use lock::*;
+pub use unlock::*;
